# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  userId: string;
  email: string;
  token: string;
}

export interface ScanPlantBody {
  imageBase64: string;
  userId: string;
}

export interface NutritionItem {
  name: string;
  amount: string;
  percentage: string;
}

export interface PlantReport {
  id: string;
  name: string;
  scientificName: string;
  description: string;
  benefits: string;
  care: string;
  nutrition: NutritionItem[];
  imageBase64?: string;
  createdAt: string;
}

export interface AddToGardenBody {
  userId: string;
  plantName: string;
  scientificName: string;
  description: string;
  benefits: string;
  care: string;
  nutrition: NutritionItem[];
  imageBase64?: string;
}

export interface SavedPlant {
  id: string;
  userId: string;
  plantName: string;
  scientificName: string;
  description: string;
  benefits: string;
  care: string;
  nutrition: NutritionItem[];
  imageBase64?: string;
  savedAt: string;
}

export interface ScanRecord {
  id: string;
  userId: string;
  plantName: string;
  scientificName: string;
  imageBase64?: string;
  createdAt: string;
}

export type GetGardenParams = {
  userId: string;
};

export type GetFavoritesParams = {
  userId: string;
};

export type GetScansParams = {
  userId: string;
};