# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Scan a plant image with AI
 */
export const ScanPlantBody = zod.object({
  imageBase64: zod.string(),
  userId: zod.string(),
});

export const ScanPlantResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  scientificName: zod.string(),
  description: zod.string(),
  benefits: zod.string(),
  care: zod.string(),
  nutrition: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.string(),
      percentage: zod.string(),
    }),
  ),
  imageBase64: zod.string().optional(),
  createdAt: zod.string(),
});

/**
 * @summary Get user's garden
 */
export const GetGardenQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetGardenResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  plantName: zod.string(),
  scientificName: zod.string(),
  description: zod.string(),
  benefits: zod.string(),
  care: zod.string(),
  nutrition: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.string(),
      percentage: zod.string(),
    }),
  ),
  imageBase64: zod.string().optional(),
  savedAt: zod.string(),
});
export const GetGardenResponse = zod.array(GetGardenResponseItem);

/**
 * @summary Add plant to garden
 */
export const AddToGardenBody = zod.object({
  userId: zod.string(),
  plantName: zod.string(),
  scientificName: zod.string(),
  description: zod.string(),
  benefits: zod.string(),
  care: zod.string(),
  nutrition: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.string(),
      percentage: zod.string(),
    }),
  ),
  imageBase64: zod.string().optional(),
});

/**
 * @summary Remove plant from garden
 */
export const RemoveFromGardenParams = zod.object({
  id: zod.coerce.string(),
});

export const RemoveFromGardenResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get user's favorites
 */
export const GetFavoritesQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetFavoritesResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  plantName: zod.string(),
  scientificName: zod.string(),
  description: zod.string(),
  benefits: zod.string(),
  care: zod.string(),
  nutrition: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.string(),
      percentage: zod.string(),
    }),
  ),
  imageBase64: zod.string().optional(),
  savedAt: zod.string(),
});
export const GetFavoritesResponse = zod.array(GetFavoritesResponseItem);

/**
 * @summary Add plant to favorites
 */
export const AddToFavoritesBody = zod.object({
  userId: zod.string(),
  plantName: zod.string(),
  scientificName: zod.string(),
  description: zod.string(),
  benefits: zod.string(),
  care: zod.string(),
  nutrition: zod.array(
    zod.object({
      name: zod.string(),
      amount: zod.string(),
      percentage: zod.string(),
    }),
  ),
  imageBase64: zod.string().optional(),
});

/**
 * @summary Remove plant from favorites
 */
export const RemoveFromFavoritesParams = zod.object({
  id: zod.coerce.string(),
});

export const RemoveFromFavoritesResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get user's recent scans
 */
export const GetScansQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetScansResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  plantName: zod.string(),
  scientificName: zod.string(),
  imageBase64: zod.string().optional(),
  createdAt: zod.string(),
});
export const GetScansResponse = zod.array(GetScansResponseItem);

/**
 * @summary Register a user
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login a user
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  userId: zod.string(),
  email: zod.string(),
  token: zod.string(),
});