# lib/api-zod/src/generated/types/addToGardenBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */
import type { NutritionItem } from "./nutritionItem";

export interface AddToGardenBody {
  userId: string;
  plantName: string;
  scientificName: string;
  description: string;
  benefits: string;
  care: string;
  nutrition: NutritionItem[];
  imageBase64?: string;
}