# lib/api-zod/src/generated/types/nutritionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */

export interface NutritionItem {
  name: string;
  amount: string;
  percentage: string;
}