# lib/api-zod/src/generated/types/scanPlantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanPlantBody {
  imageBase64: string;
  userId: string;
}