# lib/api-zod/src/generated/types/scanRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GreenScope AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanRecord {
  id: string;
  userId: string;
  plantName: string;
  scientificName: string;
  imageBase64?: string;
  createdAt: string;
}